requirements.txt
pytest
marshmallow==3.0.0rc6
PyHamcrest
pytest-pythonpath